[refactor/feat] frot - 현재 로그인된 유저의 토큰이 제대로 가져와지지 않는 문제를 해결하기위해 api moduel(@/apis/index.js)을 만들어 모든 axios들에대해 API_URL과 로그인시 받아오는 token값 을 넘겨주게 수정하였다. / 자신의 게시글과 댓글만 삭제 가능하게 구현, 댓글삭제시 window.comfirm창 띄워 확인시키기구현, data table (width값) CSS수정

+++ bankweb-pjt/src/views/DetailView.vue
@@ -1,25 +1,24 @@
 <template>
   <div class="d-flex justify-center align-center flex-column ga-6">
     <v-card class="px-13 py-12" width="950px" min-height="530px" rounded="xl">
-
       <!-- 뒤로가기 버튼 -->
       <v-btn variant="flat" @click="goBack" class="mb-4">
         <svg-icon type="mdi" :path="mdiArrowLeft" width="20px"></svg-icon>
         뒤로가기
       </v-btn>
 
-        <!-- 프로필, 아이디 -->
-        <div class="d-flex justify-space-between">
-            <div div class="d-flex align-center">
-              <img :src="articleStore.articledetail.user.profile_img" alt="프로필 이미지" class="profile-img" />
-              <p class="nickname">
-                  {{ articleStore.articledetail.user.nickname }}
-              </p>
-            </div>
-            <div class="article-category">
-              <p>{{ articleStore.articledetail.category }}</p>
-            </div>
+      <!-- 프로필, 아이디 -->
+      <div class="d-flex justify-space-between">
+        <div div class="d-flex align-center">
+          <img :src="articleStore.articledetail.user.profile_img" alt="프로필 이미지" class="profile-img" />
+          <p class="nickname">
+            {{ articleStore.articledetail.user.nickname }}
+          </p>
         </div>
+        <div class="article-category">
+          <p>{{ articleStore.articledetail.category }}</p>
+        </div>
+      </div>
 
       <!-- 글 제목 -->
       <p class="article-title">{{ articleStore.articledetail.title }}</p>
@@ -29,20 +28,20 @@
         <div>
           <p class="article-createdat">{{ formattedCreatedAt }}</p>
         </div>
-        <div class="d-flex ga-6 article-update-delete mr-8">
-            <a href="#" class="article-update">수정</a>
-            <a href="#" class="article-delete" @click="deleteArticle(articleStore.articledetail.id)">삭제</a>
+        <div v-if="accountStore.userinfo.nickname == articleStore.articledetail.user.nickname" class="d-flex ga-6 article-update-delete mr-8">
+          <a href="#" class="article-update">수정</a>
+          <a href="#" class="article-delete" @click="deleteArticle(articleStore.articledetail.id)">삭제</a>
         </div>
       </div>
 
       <!-- 글 내용 -->
-      <textarea class="article-content" readonly  >
+      <textarea class="article-content" readonly>
         {{ articleStore.articledetail.content }}
       </textarea>
     </v-card>
 
     <!-- 댓글 창 -->
-    <Comment/>
+    <Comment />
   </div>
 </template>
 
@@ -51,34 +50,40 @@ import { useArticleStore } from "@/stores/article";
 import { computed, onMounted } from "vue";
 import { useRoute } from "vue-router";
 // pictogrammers mdi icon
-import SvgIcon from '@jamescoyle/vue-icon';
-import { mdiArrowLeft } from '@mdi/js';
+import SvgIcon from "@jamescoyle/vue-icon";
+import { mdiArrowLeft } from "@mdi/js";
 
 import Comment from "@/components/Comment.vue";
+import { useProfileStore } from "@/stores/profile";
+import { useAccountStore } from "@/stores/account";
 
 const articleStore = useArticleStore();
+const profileStore = useProfileStore();
+const accountStore = useAccountStore();
+
 const route = useRoute();
 
-
 onMounted(() => {
-    const payload = {
-        articleid: route.params.id
-    }
-    articleStore.getArticleDetail(payload)
-})
+  const payload = {
+    articleid: route.params.id,
+  };
+  articleStore.getArticleDetail(payload);
+});
 
 // 게시글 삭제 함수
 const deleteArticle = function (articleid) {
-  const payload = {
-    articleid: articleid
+  if (window.confirm("정말로 게시글을 삭제하시겠습니까?")) {
+    const payload = {
+      articleid: articleid,
+    };
+    articleStore.deleteArticle(payload);
   }
-  articleStore.deleteArticle(payload)
-}
+};
 
 // 뒤로가기 함수
 const goBack = () => {
   window.history.back(); // 브라우저의 뒤로가기
-}
+};
 
 // 날짜 포맷팅
 const formattedCreatedAt = computed(() => {
@@ -88,20 +93,18 @@ const formattedCreatedAt = computed(() => {
   const date = new Date(createdAt);
 
   const options = {
-    year: 'numeric',
-    month: 'long',  // 월을 'long' 형식으로 (예: 11월)
-    day: 'numeric',
-    hour: '2-digit',
-    minute: '2-digit',
+    year: "numeric",
+    month: "long", // 월을 'long' 형식으로 (예: 11월)
+    day: "numeric",
+    hour: "2-digit",
+    minute: "2-digit",
   };
 
-  const formattedDate = new Intl.DateTimeFormat('ko-KR', options).format(date);
-  
+  const formattedDate = new Intl.DateTimeFormat("ko-KR", options).format(date);
+
   // "오전" / "오후"를 제거하고 원하시는 형식에 맞게 수정
-  return formattedDate.replace('오전', '').replace('오후', '').trim();
+  return formattedDate.replace("오전", "").replace("오후", "").trim();
 });
-
-
 </script>
 
 <style scoped>
@@ -109,47 +112,46 @@ const formattedCreatedAt = computed(() => {
   width: 44px;
 }
 
-.nickname{
+.nickname {
   font-size: 16px;
-  margin-left:16px;
+  margin-left: 16px;
 }
 
-.article-category{
+.article-category {
   display: flex;
   justify-content: center;
   align-items: center;
   border-radius: 1cap;
-  width:100px;
-  height:36px;
-  background-color:#0B5BCB;
+  width: 100px;
+  height: 36px;
+  background-color: #0b5bcb;
   color: white;
 }
-.article-title{
+.article-title {
   margin-top: 20px;
   margin-bottom: 12px;
   font-size: 24px;
   font-weight: bolder;
 }
 
-.article-createdat{
+.article-createdat {
   font-size: 13px;
-  color: #7D7D7D;
-  
+  color: #7d7d7d;
 }
 
-.article-update-delete{
+.article-update-delete {
   font-size: 15px;
 }
 
-.article-update{
-  color: #4F4F4F;
+.article-update {
+  color: #4f4f4f;
 }
 
-.article-delete{
-  color: #DE3C3C;
+.article-delete {
+  color: #de3c3c;
 }
 
-.article-content{
+.article-content {
   font-size: 18px;
   width: 100%;
   min-height: 217px;
@@ -160,5 +162,4 @@ const formattedCreatedAt = computed(() => {
   border-style: solid;
   border-radius: 1ch;
 }
-
 </style>
